modified length to floats

+++ db/migrate/20180717150925_create_trails.rb
@@ -2,7 +2,7 @@ class CreateTrails < ActiveRecord::Migration[5.2]
   def change
     create_table :trails do |t|
       t.string :name
-      t.integer :length
+      t.float :length
       
       t.timestamps
     end
